+++ src/steps/Microsoft.DotNet.UpgradeAssistant.Steps.Solution/FinalizeSolutionStep.cs
@@ -23,12 +23,13 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
 
         public override string Title => "Finalize upgrade";
 
-        public override string Description => "All projects have been upgraded. Please review any changes and test accordingly.";
+        public override string Description => "All projects have been upgraded as much as the tool is capable at the moment. Please review any changes and test accordingly. By finalizing the solution, any state tracked by Upgrade Assistant will be removed and future sessions will recalculate any potential progress that has been made.";
 
         public override string Id => WellKnownStepIds.FinalizeSolutionStepId;
 
         protected override Task<UpgradeStepApplyResult> ApplyImplAsync(IUpgradeContext context, CancellationToken token)
         {
+            // Finalizing will remove any progress tracked by Upgrade Assistant.
             context.IsComplete = true;
             context.EntryPoints = Enumerable.Empty<IProject>();
 
@@ -37,13 +38,13 @@ namespace Microsoft.DotNet.UpgradeAssistant.Steps.Solution
 
         protected override Task<UpgradeStepInitializeResult> InitializeImplAsync(IUpgradeContext context, CancellationToken token)
         {
-            if (!context.EntryPoints.Any())
+            if (context.EntryPoints.Any())
             {
-                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Complete, "Upgrade completed", BuildBreakRisk.None));
+                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Incomplete, "Finalize upgrade and delete .upgrade-assistant file", BuildBreakRisk.None));
             }
             else
             {
-                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Incomplete, "Finalize upgrade", BuildBreakRisk.None));
+                return Task.FromResult(new UpgradeStepInitializeResult(UpgradeStepStatus.Complete, "Upgrade completed", BuildBreakRisk.None));
             }
         }