+++ tests/extensions/default/analyzers/Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers.Test/AnalyzerTests.cs
@@ -7,10 +7,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
-using Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers.Test;
 using Xunit;
 
-namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.Test
+namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.Analyzers.Test
 {
     [Collection(AnalyzerTestCollection.Name)]
     public class AnalyzerTests
@@ -33,8 +32,8 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 {
                     new ExpectedDiagnostic("UA0002", new TextSpan(121, 11)),
                     new ExpectedDiagnostic("UA0002", new TextSpan(171, 10)),
-                    new ExpectedDiagnostic("UA0002", new TextSpan(307, 10)),
-                    new ExpectedDiagnostic("UA0002", new TextSpan(375, 13)),
+                    new ExpectedDiagnostic("UA0002", new TextSpan(296, 21)),
+                    new ExpectedDiagnostic("UA0002", new TextSpan(360, 28)),
                     new ExpectedDiagnostic("UA0002", new TextSpan(434, 13)),
                     new ExpectedDiagnostic("UA0002", new TextSpan(486, 13))
                 }
@@ -43,12 +42,12 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 "UA0003",
                 new[]
                 {
-                    new ExpectedDiagnostic("UA0003", new TextSpan(248, 10)),
-                    new ExpectedDiagnostic("UA0003", new TextSpan(339, 14)),
+                    new ExpectedDiagnostic("UA0003", new TextSpan(233, 25)),
+                    new ExpectedDiagnostic("UA0003", new TextSpan(335, 18)),
                     new ExpectedDiagnostic("UA0003", new TextSpan(375, 14)),
                     new ExpectedDiagnostic("UA0003", new TextSpan(416, 12)),
-                    new ExpectedDiagnostic("UA0003", new TextSpan(485, 12)),
-                    new ExpectedDiagnostic("UA0003", new TextSpan(521, 10)),
+                    new ExpectedDiagnostic("UA0003", new TextSpan(470, 27)),
+                    new ExpectedDiagnostic("UA0003", new TextSpan(506, 25)),
                     new ExpectedDiagnostic("UA0003", new TextSpan(556, 14)),
                     new ExpectedDiagnostic("UA0003", new TextSpan(605, 18))
                 }
@@ -57,13 +56,13 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 "UA0004",
                 new[]
                 {
-                    new ExpectedDiagnostic("UA0004", new TextSpan(97, 13)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(171, 21)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(336, 22)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(402, 22)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(434, 22)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(624, 21)),
-                    new ExpectedDiagnostic("UA0004", new TextSpan(716, 22)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(82, 28)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(156, 36)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(321, 37)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(398, 26)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(430, 26)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(609, 36)),
+                    new ExpectedDiagnostic("UA0004", new TextSpan(701, 37)),
                     new ExpectedDiagnostic("UA0004", new TextSpan(866, 13)),
                     new ExpectedDiagnostic("UA0004", new TextSpan(941, 21))
                 }
@@ -92,15 +91,15 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 "UA0007",
                 new[]
                 {
-                    new ExpectedDiagnostic("UA0007", new TextSpan(131, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(314, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(377, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(417, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(597, 10)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(116, 25)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(310, 14)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(362, 25)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(402, 25)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(593, 14)),
                     new ExpectedDiagnostic("UA0007", new TextSpan(617, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(675, 10)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(660, 25)),
                     new ExpectedDiagnostic("UA0007", new TextSpan(735, 10)),
-                    new ExpectedDiagnostic("UA0007", new TextSpan(826, 10)),
+                    new ExpectedDiagnostic("UA0007", new TextSpan(811, 25)),
                     new ExpectedDiagnostic("UA0007", new TextSpan(871, 10)),
                 }
             },
@@ -108,12 +107,12 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 "UA0008",
                 new[]
                 {
-                    new ExpectedDiagnostic("UA0008", new TextSpan(79, 9)),
+                    new ExpectedDiagnostic("UA0008", new TextSpan(64, 24)),
                     new ExpectedDiagnostic("UA0008", new TextSpan(112, 9)),
-                    new ExpectedDiagnostic("UA0008", new TextSpan(150, 9)),
-                    new ExpectedDiagnostic("UA0008", new TextSpan(191, 9)),
+                    new ExpectedDiagnostic("UA0008", new TextSpan(135, 24)),
+                    new ExpectedDiagnostic("UA0008", new TextSpan(187, 13)),
                     new ExpectedDiagnostic("UA0008", new TextSpan(287, 9)),
-                    new ExpectedDiagnostic("UA0008", new TextSpan(331, 9)),
+                    new ExpectedDiagnostic("UA0008", new TextSpan(316, 24)),
                 }
             },
             {
@@ -121,10 +120,10 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                 new[]
                 {
                     new ExpectedDiagnostic("UA0009", new TextSpan(102, 12)),
-                    new ExpectedDiagnostic("UA0009", new TextSpan(143, 12)),
-                    new ExpectedDiagnostic("UA0009", new TextSpan(216, 12)),
+                    new ExpectedDiagnostic("UA0009", new TextSpan(123, 32)),
+                    new ExpectedDiagnostic("UA0009", new TextSpan(196, 32)),
                     new ExpectedDiagnostic("UA0009", new TextSpan(243, 12)),
-                    new ExpectedDiagnostic("UA0009", new TextSpan(311, 12)),
+                    new ExpectedDiagnostic("UA0009", new TextSpan(307, 16)),
                 }
             },
             {
@@ -143,9 +142,27 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
                     new ExpectedDiagnostic("UA0012", new TextSpan(2249, 28)),
                     new ExpectedDiagnostic("UA0012", new TextSpan(3162, 28)),
                     new ExpectedDiagnostic("UA0012", new TextSpan(4089, 39)),
-                    new ExpectedDiagnostic("UA0012", new TextSpan(4943, 39))
+                    new ExpectedDiagnostic("UA0012", new TextSpan(4943, 39)),
                 }
-            }
+            },
+            {
+                "UA0013",
+                new[]
+                {
+                    new ExpectedDiagnostic("UA0013", new TextSpan(166, 13)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(594, 29)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(933, 14)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(1009, 25)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(1058, 10)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(1081, 13)),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(1139, 25)),
+
+                    new ExpectedDiagnostic("UA0013", new TextSpan(177, 13), Language.VisualBasic),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(463, 29), Language.VisualBasic),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(982, 10), Language.VisualBasic),
+                    new ExpectedDiagnostic("UA0013", new TextSpan(1265, 25), Language.VisualBasic),
+                }
+            },
         };
 
         // No diagnostics expected to show up
@@ -171,12 +188,23 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
         [InlineData("UA0009")]
         [InlineData("UA0010")]
         [InlineData("UA0012")]
+        [InlineData("UA0013")]
         [Theory]
         public async Task UpgradeAnalyzers(string diagnosticId)
         {
-            var diagnostics = await TestHelper.GetDiagnosticsAsync($"{diagnosticId}.cs", diagnosticId).ConfigureAwait(false);
+            foreach (var language in new[] { Language.CSharp, Language.VisualBasic })
+            {
+                var expectedDiagnostics = ExpectedDiagnostics[diagnosticId].Where(diagnostics => diagnostics.Language == language);
+                if (!expectedDiagnostics.Any())
+                {
+                    // nothing to see here, move along
+                    continue;
+                }
 
-            AssertDiagnosticsCorrect(diagnostics, ExpectedDiagnostics[diagnosticId]);
+                var fileExtension = language.GetFileExtension();
+                var diagnostics = await TestHelper.GetDiagnosticsAsync(language, $"{diagnosticId}.{fileExtension}", diagnosticId).ConfigureAwait(false);
+                AssertDiagnosticsCorrect(diagnostics, expectedDiagnostics);
+            }
         }
 
         [InlineData("UA0001")]
@@ -190,26 +218,41 @@ namespace Microsoft.DotNet.UpgradeAssistant.Extensions.Default.CSharp.Analyzers.
         [InlineData("UA0009")]
         [InlineData("UA0010")]
         [InlineData("UA0012")]
+        [InlineData("UA0013")]
         [Theory]
         public async Task UpgradeCodeFixer(string diagnosticId)
         {
-            var fixedSource = await TestHelper.FixSourceAsync($"{diagnosticId}.cs", diagnosticId).ConfigureAwait(false);
-            var expectedSource = await TestHelper.GetSourceAsync($"{diagnosticId}.Fixed.cs").ConfigureAwait(false);
+            foreach (var language in new[] { Language.CSharp, Language.VisualBasic })
+            {
+                var expectedDiagnostics = ExpectedDiagnostics[diagnosticId].Where(diagnostics => diagnostics.Language == language);
+                if (!expectedDiagnostics.Any())
+                {
+                    // nothing to see here, move along
+                    continue;
+                }
 
-            Assert.NotNull(expectedSource);
+                var extension = language.GetFileExtension();
+                var fixedSource = await TestHelper.FixSourceAsync(language, $"{diagnosticId}.{extension}", diagnosticId).ConfigureAwait(false);
+                var expectedSource = await TestHelper.GetSourceAsync(language, $"{diagnosticId}.Fixed.{extension}").ConfigureAwait(false);
 
-            var expectedText = (await expectedSource!.GetTextAsync().ConfigureAwait(false)).ToString();
-            var fixedText = (await fixedSource.GetTextAsync().ConfigureAwait(false)).ToString();
-            Assert.Equal(expectedText, fixedText);
+                Assert.NotNull(expectedSource);
+
+                var expectedText = (await expectedSource!.GetTextAsync().ConfigureAwait(false)).ToString();
+                var fixedText = (await fixedSource.GetTextAsync().ConfigureAwait(false)).ToString();
+                Assert.Equal(expectedText, fixedText);
+            }
         }
 
-        private static void AssertDiagnosticsCorrect(IEnumerable<Diagnostic> diagnostics, ExpectedDiagnostic[] expectedDiagnostics)
+        private static void AssertDiagnosticsCorrect(IEnumerable<Diagnostic> diagnostics, IEnumerable<ExpectedDiagnostic> expectedDiagnostics)
         {
-            Assert.Equal(expectedDiagnostics.Length, diagnostics.Count());
+            Assert.Equal(expectedDiagnostics.Count(), diagnostics.Count());
             var count = 0;
             foreach (var d in diagnostics.OrderBy(d => d.Location.SourceSpan.Start))
             {
-                Assert.True(expectedDiagnostics[count++].Matches(d), $"Expected diagnostic {count} to be at {expectedDiagnostics[count - 1].SourceSpan}; actually at {d.Location.SourceSpan}");
+                var expected = $"{expectedDiagnostics.ElementAt(count).SourceSpan}";
+                var actual = $"{d.Location.SourceSpan}";
+                Assert.True(expectedDiagnostics.ElementAt(count).Matches(d), $"Expected {expectedDiagnostics.ElementAt(count).Language} diagnostic {count} to be at {expectedDiagnostics.ElementAt(count).SourceSpan}; actually at {d.Location.SourceSpan}");
+                count++;
             }
         }
     }