+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the .NET Upgrade Assistant will be documented in this fil
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Current
+
+### Fixed
+- Added analyzer and code fixer to upgrade System.Web.Mvc.Controller to Microsoft.AspNetCore.Mvc.Controller. Updated ApiControll upgrade code fix to upgrade to ControllerBase, instead. [#534](https://github.com/dotnet/upgrade-assistant/pull/534)
+- VB projects that have a MyType property that requires Windows will now default to net5.0-windows [#529](https://github.com/dotnet/upgrade-assistant/pull/529)
+- Restores are now more likely to be performed if needed so errors about finding targets won't be surfaced. A clearer message will be surfaced as well if this occurs. [#525](https://github.com/dotnet/upgrade-assistant/pull/525)
+
 ## Version 0.2.226201 - 2021-05-12 ([Link](https://www.nuget.org/packages/upgrade-assistant/0.2.226201))
 
 ### Added